+++ CHANGELOG.md
@@ -17,10 +17,12 @@ Exceptions are acceptable depending on the circumstances (critical bug fixes tha
 
 ## [Unreleased]
 
+## [2.0.0] - 2025-01-17
+
 ### Changed
 
 - **BREAKING CHANGE:** corrected import state using two parts in the ID to guarantee the resource consistency
-- corrected validation inside the provider to avoid having empty URL when it's required
+- corrected validation inside the provider to avoid having an empty URL when it's required
 
 ## [1.2.1] - 2024-12-09